// src/frontend/src/client/models/CurrencyOut.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CurrencyOut = {
    base_code: string;
    target_code: string;
    timestamp: string;
    rate: number;
    symbol: string;
    stripe_supported?: boolean;
    stripe_zero_decimal?: boolean;
    stripe_no_fraction?: boolean;
    id: number;
};